v0.1a of the java client. Relies on a few dependencies because ~~JAVA~~
Specifically, because Java doesn't really allow for the same
compilation-translation-linking unit steps that C++ does, the only real way
for us to distribute a library that someone implements (without
****explicitly**** placing a restriction on the filename for) is to find a
class that implements our interface during runtime, and then instantiate it on
the fly. Turns out java doesn't make this easy, so we leverage the Reflections
library to do this quickly.